fix: refactor timeunit
