+++ crates/proof-of-sql/src/base/commitment/column_commitment_metadata.rs
@@ -169,7 +169,7 @@ mod tests {
         database::OwnedColumn,
         math::decimal::Precision,
         scalar::Curve25519Scalar,
-        time::{timestamp::PoSQLTimeUnit, timezone::PoSQLTimeZone},
+        time::{timeunit::PoSQLTimeUnit, timezone::PoSQLTimeZone},
     };
 
     #[test]
