+++ crates/proof-of-sql/src/base/time/mod.rs
@@ -1,4 +1,4 @@
 /// Native timestamp type for proof-of-sql
-pub mod timestamp;
+pub mod timeunit;
 /// typed timezone type for proof-of-sql
 pub mod timezone;
